Refactor game configuration by removing unused screen size constants; update entity count retrieval in DebugDataService; add EntityCount property to EntityRegister; introduce PlantTemplates class with growth requirements for various plants.
