+++ GameEngine/Systems/VegetationSystem.cs
@@ -5,61 +5,6 @@ using WebPeli.GameEngine.World.WorldData;
 
 namespace WebPeli.GameEngine.Systems;
 
-// Lifecycle of plants
-// seed -> growing -> mature -> reproducing -> dying
-
-
-
-// Statemachine
-// Use Plantstatus enum to determine what the tree is doing and what it should do next or what it can do next
-
-// Things we need to know about a tree
-// - Age - Affects growth and yield
-// - Health - Maybe as hitpoints to fell the tree, health system will handle this
-// - Water -Moisture from mapmanager
-// - Nutrients -We can asume that if tile material is good, the tree has nutrients
-// - Sunlight
-// - Temperature -Seasons can affect this
-// - Height -Affect mainly on rendering, age * times species or something
-// - Width - 4x4 tiles max
-// - Leaves - how much sun is blocked
-// - Branches - how much sun is blocked
-// - Roots - Dont care
-// - Fruits - Yield/loot
-// - Seeds - Purely virtual, might be in solely in tree system
-// - Flowers - Yield/loot
-// - Bark -  Yield/loot, damage resistance? or do we care?
-// - Sap - Yield/loot - Maybe
-// - Resin - Yield/loot - Maybe
-// - Wood - Yield/loot
-
-// FSM
-
-// States
-// - Seed - Can only grow 
-// - Seedling - Can only grow
-// - Sapling - Can only grow
-
-// Youth states can be stored together, we need just current state and age and tree type
-// Update tree state based on age and tree type
-
-// - Young - Grow, produce seeds, flowers, fruits
-// - Mature - Grow, produce seeds, flowers, fruits, bark, sap, resin
-// - Flowering state -> Fruiting state -> Harvesting state, Or drop fruits to reproduce
-//   - Triggers - Age, season, health, sunlight, temperature, goes to cooldown state after
-
-// - Reproducing state - Drop seeds
-// - Cooldown state - Do nothing
-
-// -Idle state - Do nothing, and remain so until neighbour tree wakes it up, or event triggers it
-
-// Addional systems needed:
-// - Weather system - Affects temperature, sunlight, water
-// - Season system - Affects temperature, sunlight, water, Already implemented
-// - Appearance system - Affects rendering
-
-
-
 public class VegetationSystem(ILogger<VegetationSystem> logger, PlantFSM plantFSM, Dictionary<Plant, PlantRequirements> plantTemplates) : BaseManager
 {
     private readonly ILogger<VegetationSystem> _logger = logger;
@@ -290,205 +235,4 @@ public readonly record struct GrowthRequirements
     public required int MinMoisture { get; init; }
     public required int MaxMoisture { get; init; }
     public required Season[] ValidSeasons { get; init; }  // When growth is possible
-}
-
-public readonly record struct PlantInstance
-{
-    // Core identifiers
-    public required int EntityId { get; init; }
-    public required Plant Type { get; init; }
-    public required Position Position { get; init; }
-    
-    // Current state
-    public required PlantMaturityStatus Status { get; init; }
-    public required int Age { get; init; }  // In days
-    public required int DaysInCurrentStage { get; init; }
-}
-
-public static class PlantTemplates 
-{
-    // Example template for grass (ground cover)
-    public static readonly PlantRequirements Grass = new()
-    {
-        IsGroundCover = true,
-        SuitableGrowthTiles = [TileMaterial.Dirt, TileMaterial.Mud],
-        MaxHeight = 0,  // Ground cover
-        MaxWidth = 0,   // Ground cover
-        GrowthStages = new()
-        {
-            { PlantMaturityStatus.Seed, new GrowthRequirements 
-                {
-                    MinDaysInStage = 2,
-                    MinTemperature = 5,
-                    MaxTemperature = 35,
-                    MinMoisture = 20,
-                    MaxMoisture = 80,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            },
-            { PlantMaturityStatus.Mature, new GrowthRequirements 
-                {
-                    MinDaysInStage = 0,  // No limit
-                    MinTemperature = 0,
-                    MaxTemperature = 40,
-                    MinMoisture = 10,
-                    MaxMoisture = 90,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            }
-        },
-        HarvestYields = new HarvestTable() // Define grass yields
-    };
-
-    public static readonly PlantRequirements Weed = new()
-    {
-        IsGroundCover = true,
-        SuitableGrowthTiles = [TileMaterial.Dirt, TileMaterial.Mud],
-        MaxHeight = 0,  // Ground cover
-        MaxWidth = 0,   // Ground cover
-        GrowthStages = new()
-        {
-            { PlantMaturityStatus.Seed, new GrowthRequirements 
-                {
-                    MinDaysInStage = 2,
-                    MinTemperature = 5,
-                    MaxTemperature = 35,
-                    MinMoisture = 20,
-                    MaxMoisture = 80,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            },
-            { PlantMaturityStatus.Mature, new GrowthRequirements 
-                {
-                    MinDaysInStage = 0,  // No limit
-                    MinTemperature = 0,
-                    MaxTemperature = 40,
-                    MinMoisture = 10,
-                    MaxMoisture = 90,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            }
-        },
-        HarvestYields = new HarvestTable() // Define grass yields
-    };
-
-    public static readonly PlantRequirements Flower = new()
-    {
-        IsGroundCover = true,
-        SuitableGrowthTiles = [TileMaterial.Dirt, TileMaterial.Mud],
-        MaxHeight = 0,  // Ground cover
-        MaxWidth = 0,   // Ground cover
-        GrowthStages = new()
-        {
-            { PlantMaturityStatus.Seed, new GrowthRequirements 
-                {
-                    MinDaysInStage = 2,
-                    MinTemperature = 5,
-                    MaxTemperature = 35,
-                    MinMoisture = 20,
-                    MaxMoisture = 80,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            },
-            { PlantMaturityStatus.Mature, new GrowthRequirements 
-                {
-                    MinDaysInStage = 0,  // No limit
-                    MinTemperature = 0,
-                    MaxTemperature = 40,
-                    MinMoisture = 10,
-                    MaxMoisture = 90,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            }
-        },
-        HarvestYields = new HarvestTable() // Define grass yields
-    };
-    // Example template for an oak tree (physical plant)
-    public static readonly PlantRequirements OakTree = new()
-    {
-        IsGroundCover = false,
-        SuitableGrowthTiles = [TileMaterial.Dirt, TileMaterial.Mud],
-        MaxHeight = 5,  // Will affect rendering
-        MaxWidth = 2,   // 2x2 tiles at full size
-        GrowthStages = new()
-        {
-            { PlantMaturityStatus.Seed, new GrowthRequirements 
-                {
-                    MinDaysInStage = 5,
-                    MinTemperature = 5,
-                    MaxTemperature = 35,
-                    MinMoisture = 30,
-                    MaxMoisture = 80,
-                    ValidSeasons = [Season.Spring]
-                }
-            },
-            { PlantMaturityStatus.Seedling, new GrowthRequirements 
-                {
-                    MinDaysInStage = 15,
-                    MinTemperature = 0,
-                    MaxTemperature = 35,
-                    MinMoisture = 20,
-                    MaxMoisture = 80,
-                    ValidSeasons = [Season.Spring, Season.Summer]
-                }
-            },
-            { PlantMaturityStatus.Sapling, new GrowthRequirements 
-                {
-                    MinDaysInStage = 30,
-                    MinTemperature = -5,
-                    MaxTemperature = 40,
-                    MinMoisture = 20,
-                    MaxMoisture = 70,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            },
-            { PlantMaturityStatus.Young, new GrowthRequirements 
-                {
-                    MinDaysInStage = 60,
-                    MinTemperature = -10,
-                    MaxTemperature = 40,
-                    MinMoisture = 20,
-                    MaxMoisture = 70,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn]
-                }
-            },
-            { PlantMaturityStatus.Mature, new GrowthRequirements 
-                {
-                    MinDaysInStage = 0,  // No limit
-                    MinTemperature = -20,
-                    MaxTemperature = 45,
-                    MinMoisture = 20,
-                    MaxMoisture = 70,
-                    ValidSeasons = [Season.Spring, Season.Summer, Season.Autumn, Season.Winter]
-                }
-            }
-        },
-        HarvestYields = new HarvestTable() // Define oak tree yields
-    };
-
-    // Method to create health component only for physical plants
-    public static HealthComponent? CreateHealthComponent(PlantRequirements template, PlantMaturityStatus status)
-    {
-        if (template.IsGroundCover)
-            return null;
-
-        var baseHealth = template.MaxWidth * template.MaxWidth * 50; // Health based on size
-        
-        var healthMultiplier = status switch
-        {
-            PlantMaturityStatus.Seed => 0.2f,
-            PlantMaturityStatus.Seedling => 0.4f,
-            PlantMaturityStatus.Sapling => 0.6f,
-            PlantMaturityStatus.Young => 0.8f,
-            PlantMaturityStatus.Mature => 1.0f,
-            _ => throw new ArgumentException($"Unknown status: {status}")
-        };
-
-        return new HealthComponent
-        {
-            Health = (int)(baseHealth * healthMultiplier),
-            MaxHealth = (int)(baseHealth * healthMultiplier),
-            RegenRate = template.MaxWidth // Bigger plants regenerate faster
-        };
-    }
 }
